requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
apispec==6.8.2
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorzero==2.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
dlib==19.24.6
executing==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.17.0
filetype==1.2.0
firebase-admin==6.3.0
flasgger==0.9.7.1
Flask==3.0.0
flask-apispec==0.11.4
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
gpiozero==2.0.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.7
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pillow_heif==0.22.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyngrok==6.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.63
rpds-py==0.25.1
rsa==4.9
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.0.0
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webargs==8.7.0
Werkzeug==3.1.3
